feat: Completely replaced the project with a luminus generated project
I'm hoping to switch entirely to clojure for the time being. Python has
proven to be an issue for a number of reasons, but mainly I've been
disapointed with the utility of SymPy. It automatically simplifies
expressions which makes it difficult to reason about use for latex
printing. This will also allow me to more freely convert between server
and client code, as clojure can run on both with minimal changes.