added login to app using Auth0
It doesn't do anything yet except display a name, but it works, and
I can now use it to start pulling protected data from the server.
I used the Single-Page-App flow from the Auth0 tutorials. Though
I would actually prefer to be using the Web Site flow so that the
app is more secure. My understanding is that the Web App flow has
the server exchange the client secret for an access token, which
simplifies the flow for the client, but I could be wrong.