Improved the style of the web page.
The unicode escaped character \u202F is a non-breaking space. React
will just display &nbsp; when provided and I don't know how to force
it through in clojurescript. The character is necessary for the
login button to align to the right with flex-box. The style is god
like. Absolutely exquisite. Some of my best work really.