feat: replaced `prime-full` with `prime-dive`, which includes the final answer
It was difficult to retrieve a final answer from prime-full, as it loses the position
of the subexpressions that need to be derived in the origional equation. So I would
get the derivation for everything, but not the full answer. I also realized that I
didn't want a bredth first approach to derivation, which is what prime-full did, as
that would result in the presented steps to solve the problem jumping all over the
place. A depth first approach will be easier for a user to follow. Luckily, prime-step
works for both approaches, so I was able to keep that code (though I needed to heavily
extend it) and just replace prime-full. I also added a `basic-derivation` problem
generator, and I'm not presenting it's results on the app. The problem generator can
take an equation of `y = f(x)` and turn it into text for the problem, text for the
solution, text for the required steps, and a set of the skills involved. This is
exactly what I've been trying to build, and now I only need to extend the functionality
a bit.