feat: added ability to select which problem to display in app
I need to grab the problem card element outside of react and update it
manually using an update event off the component. I'm also watching the
math atom for changes to update the element outside of react's render
cycle altogether. I did this because MathJax's typeset does not play
well with React's reconciliation. As time goes on, more and more or my
app code is ad-hoc, as apposed to being react based. I might ditch react
if this continues, though declarative frameworks are absolutely vital to
simplicity in application development.