fix: The exp function is now treated as a number if the input is a number
It was being treated as a function of x even when x wasn't the input, which
lead to improper detection of the skills involved and a wrong answer. Both
the exp function (e^x) and power expressions (x^5) use roughly the same code
to handle numerical inputs (e^2 or 2^5). I might want to conjoin this code
for all functions into one giant numerical input detector. This would also
come in handy once I need to implement the chain rule.