feat(app): added the option to filter problems by its required skills
There's a severe problem with re-frame and reagent where a lazy sequence won't
trigger the frameworks dependency detection. What this means is that if you
deref a reagent/atom in a `map` function, the component won't update when the
atom changes, as reagent only detects deref's that occur during the render, and
lazy sequences are only realized during reconciliation. To fix this, I add a
`let` block which deref's `filters` ahead of time, so that reagent can detect
the action. Frankly, this leads me to want to look into alternatives. I've
considered building a javascript/svelte app which interops with a clojure math
library.