feat: added fourth derivation problem and fixed issues that arose
Added variance function which identifies if a pile of math is just numbers,
like (4/3)^2, or is actually a function of a variable, like (4/x)^2. This
was necessary to decide between the product rule and the scaler rule when
the scaler is a fraction, but it came in handy for other applications.

Notably, I was able to consolidate the derivation of numbers from across the
different math function methods into the ::numeric method. This required that
I create a new defmulti key generator, distinct from the latex printing, as
latex doesn't care if math is a function or a number, but derivation does.

I also added a paren function to wrap latex strings in left-right paren's,
which will scale to their contents. There's no reason to not use these everywhere.
And I shifted more static text from prime-step methods into latex printing. The
static text is now static math literals. It's starting to look like I have a
domain-specific-language on my hands for writing latex in my system. I might
create a function for that, like `str` but that can take raw math and will auto-
matically run latex and apply dollar signs to equations.