feat: derive relative to any variable, not just x
I knew this feature was missing from the start, but wanted to wait
to implement it. Luckily, it wasn't terribly difficult to add.

I used to declare each variable as a symbole (x and y) before I was
able to use it. I removed these explicit declarations and added a
step to `convert` which makes any single letter keyword a symbol.
Technically convert is no longer pure, but I still prefer this
approach for now. Since `convert` isn't altering the inputs, and is
only affecting the hierarchy of keywords namespaced to `spacedmath.
problems`, so I'm fine with the change. It does make testing awkward,
however. They only work because the earliest tests trigger convert,
but if I reorder the tests or use a new variable without converting
they'll break. I should either consistently use convert in tests, or
add it to the base of more functions.

I reworked `variance` so that it now returns a set of variables used
as apposed to just "function" or "none". This means there's no longer
a difference between "identity" for just an x and "function" for a
function of x. It's just `#{x}`. This was built for the chain rule,
but the chain rule seems to be the only feature that requires the
distinction, so I'll just implement the distinction in the chain rule
feature specifically.

Reworked `prime-dive` and `prime-step` so that the full `[::derive]`
vector is passed into `prime-step`, as apposed to just the contents.
The vector now contains the derivations target variable, which is
relavant to `prime-step`.