feat: simplify now flattens nested associative operations
The two functions ((3 + 5) + 2) and (3 + (5 + 2)) are equivilant. My
code allows addition and multiplication to have more than two operands
to represent this fact. Though I still allowed for nested addition and
multiplication. This was fine, but I want to fix latex printing of
multiplication so that literal numbers don't look like a single number.
(5*3 was just 53, which they tell me is wrong, idk) Detecting these
issues will be easier if multiplication and addition are normalized, so
I added the feature to `simplify`. I also needed to extend latex
printing of multiplication to be able to handle a variable number of
multiply operands.