fix: support deriving variables with fraction exponents
Before this commit, the derivation of x^(1/3) would break, as the power rule expected
the exponent to be a number. Now the power rule uses `immediate-add`, which goes
through several normalization steps which should facilitate multiplication and
fractions. I built this functionality as a reusable function called `crunch-numbers`,
which I will likely try to use in the future for more features.

The term "immediate" refers to the fact that I'm not setting up an [::add...] vector,
but instead actually attempting to add the two entities right away. I have a thought,
however, that I might want to be able to retroactively configure which computations
happen immediately and which happen later. At the moment, I now have three moments
where math is resolved: there is the immediate math, resolved in the function; there
is the simplification math, which is set up by the function but resolved after the
step completes; and there is the written math, which is shown to the user as a
distinct step. I'm locking myself into which steps are done when, however, which
I might regret later. I'm realizing some steps that I display to the user might best
be handled silently, in the simplification. For example, moving the derivation past
a constant multiple—(cf(x))' = c(f(x))'. I'm also increasing the amount of code that
I might need to write. If I was able to build one function which could resolve a
`[::add...]` entity, and then use it at any stage, I could avoid having to build an
`instant-add` in addition to a simplification add. The `immediate-add` would just
throw an `[::add...]` in there and it would automatically resolve under some
conditions. I should look into this now rather than later.

I also added a `log` function, which you can wrap around any expression to print its
value without interupting computation. So `(+ (- 5 4) 3)` can just become
`(+ (log (- 5 4)) 3)`. The program won't be effected in any way besides printing "1"
to the console. It's very useful, so I'm just checking it in.