chore: remove extranious line that was used for debugging
I use vim-fireplace as a built in REPL, which allows me to run any
expression I want with the 'cpp' command. I tend to use it by adding
insignificant lines of code to my files that so that I can quickly
run my functions to see the output. This one was accidentally committed
by mistake.