fix: api now uses `\x` character literals instead of `::x` keywords
Use of keywords was awkward at best. I needed the symbols to derive
from ::symbol, but I didn't want to have to list every possible letter,
so I set up derivation during `convert` calls, but that meant that if
you didn't `(convert ::x)` before using `::x` outside a convert it would
break the api. I complained about this in a prior commit message. It was
an elusive bug waiting to happen.

Now, symbols work like numbers, in that the multimethods check for char's
and dispatch using the ::symbol key. This is much more reliable.

I also FIXED a bug where I accidentally left an explicit reference to ::x
in distrinput. Which meant that deriving with an identity would convert
all variables to x's: (e^t)' = e^x.