refactor: move the equality sign to be a part of the math-list
Before the "y" and equality was omitted by the math-list and added
by the app file. I need to allow for some problems to have a
function, however: `f(x) = x^2` as opposed to `y = x^2`. This comes
down to a new equiation, `cy^-6`, where a left variable of y makes
no since. A function specifies its independent variable, and any
other variables, c, are presumed to be constants. Right now the
system is just assuming whatever lone variable is present is the
target variable.