feat: add function definitions to mathematics
So now you can have `f(x) = x^2`, as opposed to just `y = x^2`. I needed
to add custom code to derivation printing to add the "prime": `f'(x)`,
along with rote function printing. I updated prime-step for multiplication
so that it would realize when you were multiplying by a constant variable,
`cx^2`, and it would treat it like a number. That was orthoganally
related. Updated `basic-derivation` so that it now grabs the target
variable from the function definition, `f(x)` as apposed to the first
variable found in the equation, which allows me to add constant variables.
The functions are only unary for now. I'll need to fix that if I ever get
around to multivariable calc, but there's many more things I'll have to
fix as well, so I'm not too worried about it.