chore: made the equation side panel even bigger
I desperately need to make the left panel automatically adjust to its contents,
which are static. At the moment, equations don't break, so they just overlap
with the main content if I don't get the size right, but I really don't want
to have to keep checking in these changes. I'm using flexbox, so it should be
possible to just allow the left panel to grow, but I'm not sure if that will
play nice with MathJax.