feat: derive fractions with the variable in the denominator
The expression `5/x^4` can now be derived. I did by going through
an inversion step, `5x^(-4)`, which allows the power rule to be applied.
I did this as it's likely the step that a person would follow, and
Stewarts calculus book seems to think so as well, given that this problem
is introduced before the quotient (division) rule. It required that I add
an insta-invert, which slaps a -1 exponent on some math but simplifies it
down if the math was already an exponent: `(x^5)^(-1) = x^(-5)`.

This is another example of some math which gets instantly computed before
being displayed to the user, just like insta-add, and I really think I
should build this in as a first-class feature. I'm thinking that the
derivation code shouldn't be locked into applying the math instantly, but
instead the system as a whole can be configured such that different
mathematical steps are displayed to various degrees of verbosity. Some
steps would be "silent", like insta-add and insta-invert. Some steps would
be "simplifications", which would put extra steps in the stand-alone
equations, so "math = unsimplified \r = simplified", but would get no
english prose. And the last of the steps would be "verbose", which would
get the full treatment of english prose and stand-alone equality.

This has a number of advantages.
* It would be more flexible, as I would be able to change verbosity
  retroactively.
* It would allow me to make the setting dynamic. Some users who understand
  certain steps could have those steps omitted, whereas other users who
  are just now learning those steps would still see the verbose explanations.
* Simplifications could be stacked. If I have to derive `4x^2 + 5x^3` I don't
  want to have an `=` step for the distribution of the derivation and to move
  the derivation inside the constant scaler *twice*. If the step is marked as
  a "simplification" step (I need a better name) than they can be applied all
  at once, especially if they apply to different parts of the equation.
* It would reduce the amount of coding I have to do, as both instant
  addition and simplification step addition could use the same code.