fix: decimal exponents no longer become massive fractions
The insta-add, which is used by ::power derivation to subtract one from
the exponent, was treating all numbers like fractions. Integers were
being put over 1, added with cross multiplication, and then reduced
using the greatest common divisor. With decimal numbers, however, this
was producing massive numerators and denominators. I've updated the
insta-add function to use a numerical approach if neither number is a
fraction. Even though it's weird to have two entirely different chunks
of code based on the structure of the input, this way maintains the
style of the starting problem, which is an intended feature.