feat: add support for constant exponents
The insta-add function, which is used by `::power` to subtract 1 from
exponents, had no default statement and no condition for constants. I
have added an `:else` condition which falls back to just returning an
`::add` structure with the parameters, effectively making it not an
"insta" add. This actually works really well, as a constant, like `e`,
will show up as `e - 1` (or `e + -1` in the current code). I really need
to work on my simplification system that I've talked about in prior
commits.