chore: add one more equation of note to math list
Stewart's 8th edition calc book, where I've been getting my equations,
has this one as `z = A/y^(10) + Be^y`. See the problem? There is nothing
to identify the `y` variable as the actual variable as opposed to the
`A` or the `B`, other than common sense. The variable detector at the
moment checks the function variable first, but falls back to grabbing
the first variable it can find in the equation if there isn't a left
side function. In this case the first variable is `A`, and at the moment
I would rather not program in custom code to prioritize some variables
over others, though I might do that eventually.