feat: derive identities with more nuanced inputs
A function which takes a non-linear input requires the use of the chain
rule, which I have not implemented, but a linear function still allows
for `e^(x+1)`. I have now extended my identity derivation handler to
actually distribute the given input, as opposed to just the variable.
This is also the last of Stewart's 8th calc section 3.1 problems that I
plan on implementing by hand. The remainder could be solved in multiple
different ways using later sections of the book, which ruins the skill
detection, so I'm skiping them for now. I think it's time to refactor
some stuff... tomorrow.