feat: move login flow to back end
There are several different flows for OAuth2 login. The one I was using
is a frontend flow, which is less secure, as it requires that I trust
the user to not alter the code malitiously. I have now switched to a
backend flow, or "native web" flow. Because the backend flow is
performed with code that is inaccessible to the user, I can trust that
it will run correctly. In this way, nither the user, the google servers,
nor myself have to trust each other. I also ripped out the Auth0 code,
which was proving to be difficult to manage, and replaced it with the
google api directly. I pulled in some libraries from gorilla, which seem
pretty standard, along with an http library. This was just based on some
quick googling. I was hoping to stick to core libraries, but this seems
like the approach everyone uses. Next step is to get an actual SQL
server running with go.