feat: add pattern matching prime solver
This is intended to ultimately replace the current prime-dive. I'm
calling it prime-pattern for now to differentiate (lol) the two. This is
something I've talked about in prior commits if you search, but the gist
is that I always thought I could implement all rules declaratively, as
opposed to through a multimethod. This would be more flexible, as I
could change how rules are realized retroactively. It's also likely
better for the combinatorial complexity of derivation. For example,
deriving a scaler times a function of x, `5e^x`, would require two
different rules, as it's commutative: `5e^x = e^x5`. If I implement
derivation as a pattern, however, I can extend the pattern matching to
account for commutative structures automatically, and all commutative
structures will get the functionality without any additional code. It
also allows me to change the order in which patterns are (I literally
just saw a cockroach on my wall and had to go after it) realized, so
that I can have depth first or breadth first step-by-step approaches, if
I find that necessary. Ultimately, I'm just trying to undercut the
combinatorial complexity of the code.

The most important new functionality is math-match, which identifies a
matching pattern by returning a map of variable conversions between the
pattern and the match. So a pattern of `[::sin \x]` would match `[::sin
\t]` with a map of `{\x \t}`. This was modeled after how Stewarts 8th
edition calc book describes its rules. Since he was able to distill the
rules of calculus into a form that simple, I believe I should be able to
aswell. That's a decent rule for most things. If you're trying to
determine how much information you need to retain, think back to the
mechanics of the informations source. A phone number need only retain
the exact digits and source country, as that is all the telephone
network needs to connect your call.

Most of the complexity of math-match came from the need to bail on an
attempted match whenever there's a collision between different
variables. For example, the pattern `[::mult \x \y]` requires that the
variables be different. So `[::mult \u \u]` will momentarily be
considered a match, until the second "u" is discovered, a collision will
trigger, and the match will fail. Conversely, the pattern `[::mult \x
\x]` will require that the variables be the same, which is how the
pattern `[::derive [::sin \x] \x]` knows that it should only work on a
Sine of the target variable. The expression `[::derive [::sin \u] \v]`
would have two different variables for `x`, which can't be right. (In
fact, the engine isn't calculus specific. Just the rules!) This became
particularly complicated whenever I started dealing with nested
structures. The pattern matcher actually allows for variables in the
pattern to match deep math expressions, which is useful for "constants",
that can be arbitrarily deep: `\sin(5)/2`. Sometimes these constants
have constant variables.  Any variable which isn't the derivation target
is considered a constant, so a pattern of `[::derive [::mult \c \x] \x]`
is looking for any constant of arbitrary complexity multiplied by the
target variable. The collision detection then needs to work at any
depth, otherwise `[::derive [::mult [::sin \c] \x] \x]` won't work as it
should. In the end, I wound up using `variance` to create an inverse map
from the match variables to the pattern variables, which I then use as a
second collision detection.  So a pattern of `[::mult \c \x]` will
reject `[::mult [::sin \u] \u]` but will accept `[::mult [::sin \v] \u]`
with the mapping `{\c [::sin \v] \x \u}`. This is mostly facilitated by
`clean-merge`, which rejects many types of collisions.