feat: switched to pattern matching
Well that was complex. There's so much to say about it too. I managed to
keep all but the last problem in the math list, which I decided to
disable for now. It's the `e^{x+1}` which is breaking it, as that could
almost require the chain rule. I really don't want to weaken my patterns
just for that, and I know I'll be getting to it later, so despite
Stewarts Calc book introducing it earlier then chain rule, I'm putting
it aside till then.

Adding a bug fix for latex printing of exponents on exponents:
`(x^5)^5`.

I broke my identities with extra brackets at some point, not sure when,
but I fixed that.

There's two new simplifications, which are actually just based on insta-
functions that I had before. I really do need to consolidate insta-
simplifications, post simplifications, and derivation steps into the
pattern engine. At the moment, the pattern engine can't handle math,
which is why deriving exponents, `x^2`, brings in a subtraction,
`x^(2-1)`. I might have to extend the pattern engine so that it can
crunch numbers, in addition to run matches. Granted, at that point, how
is that any different than the functions I had before? We'll see.

I had to change the pattern matching approach drastically. I'll
document it soon, so you might search for that, but I'll give a simple
explanation here as well. There are three things to keep track of:
well-defined variables, functions of variables, and constants. If a
pattern variable `u` matches directly with a math variable `x`, `u`
becomes a well defined variable. A function of `u`, (`[\f \u]`), may
match any arbitrary math, which may or may not include `u`, but can't
include any other well-defined variable. So if `v` was found to be `y`,
and the math that `f` is attempting to match has a `y` in it, the match
should fail. Finally, a pattern variable `u` may also match arbitrary
math, as long as it doesn't include any well-defined variables. That
seems to cover all my bases without requiring too much information in
the patterns themselves.

I had to extend math-match to account for associative operators, as I no
longer have code which can just treat it as such. For now, it just
splits the remaining operands into a seperate operation,
`x + y + z = x + (y + z)`, but I really want to make it more clever than
that. If the starting operator is associative, then the resultant
pattern should be associative as well. You could prove it
mathematically. I should be able to have the pattern engine recursively
realize the pattern but without the developer or the user seeing it
happeneing. Granted, the way the api of math-match works now, that
wouldn't work, so I'll have to create a bigger function which does the
full pattern matching and result construction to fit this in.

I switched the prime-dive tests over to prime-pattern so that I could
check for unintended regressions. I also added a bunch more tests
whenever I found a problem that wasn't hit. This leaves prime-dive with
no tests, but I think I'm about to start ripping it out. It's officially
legacy code.

I added a todo list, which I think we'll both find useful. How many
times have I brought up consolidating simplifications? I need somewhere
to keep those thoughts that isn't hidden deep inside some git history
that virtually no one will read.