feat: add detail-math-list with verified answers
I keep having to manually check if the math-list is being derived
correctly on the app, as the tests don't seem to be exhaustive enough to
catch every bug that the math-list can. It seems like I should just
maintain a list of answers alongside the problems, so that any issues
may be detected more efficiently.