feat: remove all deprecated prime-dive code
Patterns are the way of the future. This code is no longer used and no
longer tested. I will not being developing on it further, so I'm getting
rid of it to focus on other things. I might still scavenge the
insta- simplification functions though, but I always have the history
for that. I'll try to leave a note in that commit if I ever bring the
code back in any capacity.