feat: implement in-app equation editor
The math language in use is one I just invented based on asciimath. I
call it mafs, and it should allow me to add problems faster than typing
directly into clojure. I used instaparser as I found a user online that
said it was the best parser they'd ever used in any language, which
sounded enticing. Instaparser can handle multiple different types of
grammers, but I went with an EBNF grammer in a bnf file as instaparser
referred to EBNF as the clearest grammer available. I extended the
eq.html page with a final equation output. If you type mafs into the
textarea, the page will show you the interpreted equation printed to
latex.

I had to create a new swig macro to load the bnf file. ClojureScript
doesn't make loading text files very easy. Cljs files run on the web, so
loading files at runtime is a pain and introduces latency to the load
times of the application. Loading in plain Clojure is comparatively
easy, so I created a macro that would load a file at compile time and
replace itself with the text contents. I'm surprised this isn't a part
of Clojure core.