feat: allow user to add problems to sqlite file
I'm using the GORM object-relational model library, as it was the first
thing that came up when I googled Go ORM and has decent looking
documentation. I'm also using the built in JSON parser, so no issues
there. I did try a "sqlite3 driver," which was a collosal pain to work
with (ORM's are almost vital).

My two choices for a database were, in my opinion, Postgres and Sqlite.
Postgres is open source, free, and reliable, and I still suspect I'll be
using it to track student data, as postgres is better at multiuser data.
For the problems specifically, however, I prefer Sqlite. Sqlite runs
entirely in the querying language, unlike postgres which is a separate
server. This makes it much simpler to push to AWS or any other computer
I want. The problems really shouldn't be changing that often, too, so
the lack of comprehensive multiuser support doesn't bother me much.
Sqlite also has less comprehensive value-type options: text, int, real
(number), blob, and null. Of course, you can store other value types,
like a boolean, by casting them to a supported value type, like an int
of 0 or 1. For the problems, this is fine, as I plan on just storing
their values as string serialized json. This would be an issue if I
wanted to query the internals of the json, but anything I want to query
can just be saved off to the side. This is how I handle skills. The
clojure code just specifies all skills involved in the problem so that I
can query by skill.

Note that I'm also storing the answer beside the problem. This isn't
strictly necessary, as I could always rederive the answer on the fly,
but it helps for verifying the correctness of the system. This way the
user who adds the problem (me, usually) will be able to check the
answer, and I can add a sanity check to the system which insures that
the same answer is derived later. This will also give me a huge cache of
unit test data that may be used to check the system after I've made
changes to the derivation code, which I plan to do a lot.