chore: add links and text to help app visitors
I'm about to start sharing the link with professors, so I thought I
figured it was time to actually add some help text to the equation page.
I'd love to make the language more robust but it's just an internal tool
and this is quicker. I'll have to update the help text frequently, but I
can live with that.