fix: don't break when printing `::sin`
The way I implemented the latex multimethod is fragile. When you print a
vector, it dispatches based on the first element, but when you print a
keyword, it dispatches just the keyword. This means that `::sin` and
`[::sin]` dispatch to the same function, which assumes it will get a
vector. I had to implement a converter to a vector, but this could break
many functions. Switching to pattern matching might be a decent fix, but
then I wouldn't be able to split the multimethod into many files if I
wanted to, not that I was planning on it.

I discovered this while working with my mafs editor. While typing
"f(x)=sin(x)" I briefly landed on "f(x)=sin", which created a `::sin`
without a vector, breaking my latex code. Invalid states for mafs is
going to be too much work to test (strings have nearly infinite invalid
states) so it will make more since to wrap the parser in a try catch.

I also removed the extranious call to `name` for symbols (single
letters: \x) as it didn't actually change anything. The name of a string
is the string.