fix: print latex of lone `::exp` without breaking
A lone `::exp` gets converted to [::exp], and then the recursive call to
latex with the last item in the array is a lone `::exp`, which creates
an infinite loop. I switched to using the first of the rest of the
vector, which just comes out to being nil when the vector only has one
item. This is a clunky and unintuitive fix, which I plan to swap out
with something smarter later, probably based on the arity of the
operator itself.