chore: added tasks for simplifications engine
It's time. I've talked about the simplification engine before, but I'm
actually going to start working on it. The important parts of the system
are not the exact steps taken to solve a derivative problem, but the
skills that were used to solve it, and the final answer. Right now, the
final answer is pretty bad, and it's making the code look more fragile
by association. It looks like the system doesn't understand the math
that it's doing when you see 2*(1/2) not getting reduced. Granted, the
system is supposed to avoid some "simplifications" if they aren't
something a student would feasably want to perform. It's okay to leave
210^3 alone, as a student would probably leave it like that as well. But
2^2? Yeah, no one is writing that down. They're writing down 4.

I call this behavior anthropomorphic mathematics--it does what a person
would do. This approach to math is actually far more complex than the
alternative, which is to normalize everything down to decimal numbers
and just run the computations. The expression 2*(1/2) is really easy to
simplify if you just compute 2*0.5 and find out it's 1, but then you run
the risk of "simplifying" everything, and having weirdly complex
answers, like how 210^3 is 9261000.

In order to maintain anthropomorphism, which I'll just call anthropism
for short (from the greek work anthrōpos) I'll need to build all of my
simplifications to handle not only decimal numbers (0.5*0.5=0.25) but
also fractions (1/2*1/2=1/4) and even irrational numbers
(0.33…*0.33…=0.11…). Not to mention that a lot of these will apply even
when not dealing with numbers (a/a=1). This means we have a classic
cross-combinatorial problem on our hands. My solution is going to be to
normalize to one of the forms and convert back when I'm done. I call
this a variation bottleneck. The trick is to maintain enough information
about the origional form to be able to convert back. Fractions seem to
be my best bet for this. I can convert 0.356 to 356/1000 very easily,
but 1/3 doesn't convert into a floating point number without some loss.
Irrational numbers are harder, but ultimately they're just finite
decimal numbers with elipses. The real issue is to display the proper
amount of digits, which should be based on the input, but won't fit into
a fraction. I might have to pass that along as metadata or use a lens
and just maintain that information outside the simplification call.

Then there's the issue of symbolic simplifications: a/a=1. I think most
of my rules will be symbolic, as there are just more of them. If I have
code which knows how to add numbers together, I don't also need to teach
it that a-a=0. Clojure's add function already knows that.

None of this is to mention the timings, which aren't very well named.
There's instant--the user never sees it--afterward--the user sees it but
it's done with an equality before moving on--and belated--the
simplification is avoided except as an extra answer at the end. I'll use
instant if the simplification is super obvious, afterward if the
simplification might lose someone but is worth doing right away, and
belated for simplifications that I actually doubt most students will
perform. Doing the belated simplifications at all is just a convenience
for checking answers against the system. Some students might recognize
11^2=121 right away, and some won't. Both should be able to check their
answers easily, so I'll provide both.