feat: add latex-cost to determine latex complexity
I added this to aid in the simplification process. In general,
simplification of math expressions has the goal of creating a less busy
written form. Latex-cost creates a score based on how many glyphs a math
expression will produce when printed in a latex engine. It doesn't know
when parenthesis are necessary, however, so I added a parens math
structure for that. I'll need to create a function that can add
parenthesis based on ambiguity of the expression. Currently, I handle
that in the latex printing, but that has issues. First of all, use of
parens depends on the outer and inner expression. The expression x^2
doesn't require parens, but (2x)^2 does. This means a paren adder will
need to "look ahead," to attain enough context to do its job, which I
don't want to build into the latex printer. In addition, the latex
printer is only suppose to have one responsibility, and though it would
make since to roll parenthesis in there, there are some situations where
I will want extra parens, even if latex wouldn't find them necessary.

Normally I would try to be more clever about generating a latex score
than this. You would think I could just print the latex and then count
the length of the string, after all. But latex has a lot of extra
characters, and string length doesn't correlate to printed glyph count.
If I had any way of tapping into a latex rendering engine I could make
this more clever, but for now I'm forced to write this buisness logic
myself, as it would be far too much work to derive it.

Most of the math expressions have a predetermined score for their
operator. For example, subtraction has three (minus and two spaces).
Exec functions like sin() use their name length, for obvious reasons.
The only weird item is addition, which has a delimeter between a
variable number of operators (as it is associative and commutative).
Multiplication has a variable number of operators, but it uses no
delimeter, so it gets to use the same code as the other operators, with
a score of 0.