feat: add math-unify, more dynamic pattern matcher
Math-unify, unlike math-matcher which it replaces, is capable of
recognizing variation in commutative and associative operators. For
example, a pattern of a+0 should match 0+x, even though the positions of
the variable and the 0 are reversed, as addition is commutative.
Math-unify recognizes that possibility, and successfully matches the
pattern.

Unlike math-matcher, this means that math-unify can result in multiple
possible matches. Currently, a+b will always match 2+3 twice, as
(a,b)=(2,3) and (a,b)=(3,2). This is not desirable, and I'll need to
update math-unify to only check varying order when the pattern operators
are different. For example, a+0 should check both orders of a and 0, but
a+b shouldn't, as every match with trigger both orders.

As part of this change I copied math-match's tests and updated them to
check math-unify. The only changes I had to make to get them working was
to wrap the result in a vector, as math-unify returns all possible
matches, as opposed to one, and then I had to retrieve the first result
in the two tests that did, in fact, return multiple results.

There is one other possible issue I can see, and that is that pattern
still have to match the entire addition, as opposed to only part. So u+0
works, because the variable u will grab all remaining operands, but u-u
won't successfully match x+y-y, as it would have to match the y-y while
not matching the x. I'll need to fix this as well.

And, of course, I still have the problem that the patterns don't
understand recursion. A distribution pattern would be written like
c(a+b)=ca+cb, but really you want the pattern to allow for any amount of
operands in the addition. I can detect this recursion automatically, but
with the way math-unify currently returns the matched results as a map,
{c 1, a 2, b 3}, there's not a very clean way to handle this. If I made
a function which combined the pattern matching and replacement, such
that you hand it equalities, and it passes back updated expressions.

This is a solid step toward my improved simplification engine.