feat: switched from math-match to math-unify
I'm having to grab the first match to make this work, but for now I'm
happy with that. Math-unify may find more matches, but for prime-pattern,
which is the only place I was using math-match, using only one of the
possible matches is fine. For simplification it will be different,
because any random match might be better than others, so I might have to
check them all, but for now, math-match's one result can be emulated by
simply taking the first match and throwing the rest away.

Later on I'll have to check if the first result is actually the most
desirable, and by that I mean that the first result most closely matches
the origional structure of the math, a+b matching 1+2 should match as
(a,b)=(1,2), not the other way around. This is the best way to ensure
that I maintain the origional structure as much as possible, which helps
calculus students understand where the answer came from.