feat: add rule-simplify, a rules engine simplifier
This works, but it is slow, and has at least one known bug.

I will absolutely need to start benchmarking the rule-simplify and try
to get it to be faster. The tests used to finish nearly instantaniously;
now they're taking many seconds. As far as I can tell, this comes from
three issues.

First, one simplification may make another rule viable. For example,
factioring out `a` from 2a+3a makes the addition of 2 and 3 possible,
resulting in 5a. In order to satisfy this, the rule-simplify function
re-runs itself whenever the math changes, to see if any new rules may
get through. Once the math stops changing it returns.

Second, simplification is recursive, and this recusion happens on every
loop, regaredless of whether or not the recursion has any chance of
finding another simplification. Some recursion makes since when a new
structure is created. For example (a^b)^c becomes a^(b*c), which has a
new (b*c) structure. This should go through simplification, but if the
variable a itself has already been simplified, it should not be
recursed. The looping and recursion compound to make a very inefficient
process.

Third, matching of commutative and associative operators requires that
every possible order of operands be checked. This means that a simple
structure like a+b+c which is being matched against the pattern x+y will
be checked a handful of times, and if the structure and pattern are
nested the recursions will also trigger multiple times. I doubt this is
the main source of computation time, but I'll have to wait for the
benchmarks to be sure.

I updated mafs so that parenthesis correctly allocate operators. Now if
you write (a+b+c) it will come out as [::add \a \b \c] and if you write
(a+b)+c it will come out as [:add [::add \a \b] \c]. This was important
for the pattern matching of simplification rules, which were written in
mafs. If a pattern says (a+b)+c, that's because that structure is
important. Previously I was counting on the old simplification function
to consolidate operators, but I'd really like to get rid of that, and I
don't always want to consolidate them anymore. Parse-mafs now converts
the code directly to whatever you tell it, skipping any and all
simplifications, which is generally better from a Single Responsibility
Principle.

Currently the simplification just goes from left to right in the
pattern, with no backtracking to check if another simplification path
might have been better. This is fine for now, as the current rules are
all "no brainer's" like a+0=a, but in the future I'll want to cross
check simplifications against their latex score, just to be sure. Some
simplifications, like factoring out a common multiple, ab+ac=a(b+c), are
only simplifications if the latex gets simpler, which is only true some
of the time.